wraith-config-gen/requirements.txt
selenium
webdrivermanager
ruamel.yaml
python-slugify
